requirements.txt
asgiref==3.11.1
dj-database-url==3.0.1
Django==6.0.3
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==23.0.0
openpyxl==3.1.5
pillow==12.2.0
psycopg[binary]==3.2.12
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
